Moved EntityTypeName/Resolver out the .model.jpa package into the .model package
